vishwaas/requirements.txt
apturl==0.5.2
asgiref==3.6.0
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-cors-headers==3.13.0
django-filter==22.1
django-jsonfield==1.4.1
djangorestframework==3.14.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docx2txt==0.8
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.1
fasteners==0.14.1
Flask==2.2.3
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==2.8
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
packaging==23.1
pandas==1.5.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.0.0
protobuf==4.22.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==7.3.2
pytest-django==4.5.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.3
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
tomli==2.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==2.2.3
xkit==0.0.0
zipp==3.15.0
